Move to seperate models script file
